New data in new format with addition of relative humidity. Made changes to the code to accommodate.
If there were gaps in the time component of the dataset, lines were drawn from one point to the next regardless of a data gap. Created data frames that contained all combinations of the 4 nodes (will need to change to 5 sometime), and all of the dates (for daily, weekly, and monthly). Merged the data with these frames. Now all of the gaps are displayed properly and all four wind roses show up for every month.

The wind roses take a long time to generate, and were being called each time the app was run. When the app is deployed, it is not necessary to do this every time the app is called. I commented out the code that calls it - need to uncomment any time that the app is first deployed.

The lines in the wind graphs were too light to see. Needed to add to the wind rose code in load_barrow_data.R:
                         panel.grid.major.y = element_line(size = .5, colour = "gray80"),
                      panel.grid.major.x = element_line(size = .5, colour = "gray80"),
                      panel.background = element_rect(fill = "white")
          Also removed theme_bw(), since it was blocking changes to the panel background and grid lines.

The NA values at the top of the curve for soil moisture are due to saturation of the sensor. Perhaps just limit the graph to the saturation value and put a label that says ‘saturation limit’.